requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
absl-py=0.13.0=py39h06a4308_0
aiohttp=3.7.4.post0=py39h7f8727e_2
allennlp=2.9.0=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.5=pypi_0
async-timeout=3.0.1=py39h06a4308_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pypi_0
backports-csv=1.0.7=pypi_0
base58=2.1.1=pypi_0
beautifulsoup4=4.10.0=pypi_0
black=22.1.0=pypi_0
blas=1.0=mkl
bleach=4.1.0=pypi_0
blinker=1.4=py39h06a4308_0
blis=0.7.5=pypi_0
boto3=1.20.47=pypi_0
botocore=1.23.47=pypi_0
bottleneck=1.3.2=py39hdd57654_1
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h27cfd23_0
ca-certificates=2022.3.29=h06a4308_1
cached-path=1.0.2=pypi_0
cachetools=4.2.2=pyhd3eb1b0_0
catalogue=2.0.6=pypi_0
certifi=2021.10.8=py39h06a4308_2
cffi=1.14.6=py39h400218f_0
chardet=4.0.0=py39h06a4308_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
checklist=0.0.11=pypi_0
cheroot=8.6.0=pypi_0
cherrypy=18.6.1=pypi_0
click=8.0.3=pyhd3eb1b0_0
coverage=5.5=py39h27cfd23_2
cryptography=3.4.8=py39hd23ed53_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py39h06a4308_0
cymem=2.0.6=pypi_0
cython=0.29.24=py39h295c915_0
dataclasses=0.8=pyh6d0b6a4_7
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.4=pyhd3eb1b0_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.4=pypi_0
executing=0.8.2=pypi_0
expat=2.4.1=h2531618_2
fairscale=0.4.5=pypi_0
fastnlp=0.7.0=pypi_0
feedparser=6.0.8=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.4.2=pypi_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
fsspec=2021.8.1=pyhd3eb1b0_0
future=0.18.2=py39h06a4308_1
giflib=5.2.1=h7b6447c_0
gitdb=4.0.9=pypi_0
gitpython=3.1.26=pypi_0
glib=2.69.1=h5202010_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-api-core=2.5.0=pypi_0
google-auth=1.33.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.1=py_2
google-cloud-core=2.2.2=pypi_0
google-cloud-storage=1.44.0=pypi_0
google-crc32c=1.3.0=pypi_0
google-resumable-media=2.1.0=pypi_0
googleapis-common-protos=1.54.0=pypi_0
grpcio=1.36.1=py39h2157cd5_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=3.6.0=py39ha0f2276_0
hdf5=1.10.6=hb1b8bf9_0
hickle=3.4.5=py_0
huggingface_hub=0.2.1=pyhd3eb1b0_0
icu=58.2=he6710b0_3
idna=3.2=pyhd3eb1b0_0
importlib-metadata=4.8.1=py39h06a4308_0
importlib_metadata=4.8.1=hd3eb1b0_0
iniconfig=1.1.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.8.0=pypi_0
ipython=8.0.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.5=pypi_0
iso-639=0.4.5=pypi_0
jaraco-classes=3.2.1=pypi_0
jaraco-collections=3.5.1=pypi_0
jaraco-context=4.1.1=pypi_0
jaraco-functools=3.5.0=pypi_0
jaraco-text=3.7.0=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.0.3=pypi_0
jmespath=0.10.0=pypi_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
jsonnet=0.18.0=pypi_0
jsonschema=4.4.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.1.2=pypi_0
jupyter-console=6.4.0=pypi_0
jupyter-core=4.9.1=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.2=pypi_0
kiwisolver=1.3.1=py39h2531618_0
lame=3.100=h7b6447c_0
langcodes=3.3.0=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1d223b6_11
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.17.2=h4ff587b_1
libstdcxx-ng=11.2.0=he4da1e4_11
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
llvm-openmp=12.0.1=h4bd325d_1
lmdb=1.3.0=pypi_0
lxml=4.7.1=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py39h06a4308_0
markupsafe=2.0.1=pypi_0
matplotlib=3.4.3=py39h06a4308_0
matplotlib-base=3.4.3=py39hbbc1b5f_0
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
more-itertools=8.12.0=pypi_0
mpi=1.0=mpich
multidict=5.1.0=py39h27cfd23_2
munch=2.5.0=pypi_0
munkres=1.1.4=py_0
murmurhash=1.0.6=pypi_0
mypy-extensions=0.4.3=pypi_0
nbclient=0.5.10=pypi_0
nbconvert=6.4.1=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.3=h7f8727e_0
nest-asyncio=1.5.4=pypi_0
nettle=3.7.3=hbbd107a_1
nltk=3.6.5=pypi_0
notebook=6.4.8=pypi_0
numexpr=2.7.3=py39h22e1b3c_1
numpy=1.21.2=py39h20f2e39_0
numpy-base=1.21.2=py39h79a1101_0
oauthlib=3.1.1=pyhd3eb1b0_0
olefile=0.46=pyhd3eb1b0_0
openh264=2.1.0=hd408876_0
openssl=1.1.1n=h7f8727e_0
packaging=21.0=pyhd3eb1b0_0
pandas=1.3.4=py39h8c16a72_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pathspec=0.9.0=pypi_0
pathtools=0.1.2=pypi_0
pathy=0.6.1=pypi_0
patsy=0.5.2=py39h06a4308_0
patternfork-nosql=3.6=pypi_0
pcre=8.45=h295c915_0
pdfminer-six=20211012=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.4.0=py39h5aabda8_0
pip=21.2.4=py39h06a4308_0
platformdirs=2.4.1=pypi_0
pluggy=1.0.0=pypi_0
portend=3.1.0=pypi_0
preshed=3.0.6=pypi_0
prettytable=3.2.0=pypi_0
prometheus-client=0.13.1=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.26=pypi_0
protobuf=3.17.2=py39h295c915_0
psutil=5.9.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
py=1.11.0=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pydantic=1.8.2=pypi_0
pydeprecate=0.3.1=pyhd8ed1ab_0
pygments=2.11.2=pypi_0
pyjwt=2.1.0=py39h06a4308_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py39h2531618_6
pyrsistent=0.18.1=pypi_0
pysocks=1.7.1=py39h06a4308_0
pytest=6.2.5=pypi_0
python=3.9.7=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-docx=0.8.11=pypi_0
python_abi=3.9=1_cp39
pytorch=1.10.0=py3.9_cuda10.2_cudnn7.6.5_0
pytorch-crf=0.7.2=pypi_0
pytorch-lightning=1.4.9=pyhd8ed1ab_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=py39h7f8727e_1
pyzmq=22.3.0=pypi_0
qt=5.9.7=h5867ecd_1
qtconsole=5.2.2=pypi_0
qtpy=2.0.1=pypi_0
readline=8.1=h27cfd23_0
regex=2021.8.3=py39h7f8727e_0
request=2.83.1=0
requests=2.26.0=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
s3transfer=0.5.1=pypi_0
sacremoses=0.0.43=pyhd3eb1b0_0
scikit-learn=1.0.1=py39h51133e4_0
scipy=1.7.1=py39h292c36d_2
seaborn=0.11.2=pyhd3eb1b0_0
send2trash=1.8.0=pypi_0
sentencepiece=0.1.96=pypi_0
sentry-sdk=1.5.4=pypi_0
setuptools=58.0.4=py39h06a4308_0
sgmllib3k=1.0.0=pypi_0
shortuuid=1.0.8=pypi_0
sip=4.19.13=py39h2531618_0
six=1.16.0=pyhd3eb1b0_0
smart-open=5.2.1=pypi_0
smmap=5.0.0=pypi_0
soupsieve=2.3.1=pypi_0
spacy=3.2.1=pypi_0
spacy-legacy=3.0.8=pypi_0
spacy-loggers=1.0.1=pypi_0
sqlite=3.36.0=hc218d9a_0
srsly=2.4.2=pypi_0
stack-data=0.1.4=pypi_0
statsmodels=0.13.0=py39h7f8727e_0
tempora=5.0.1=pypi_0
tensorboard=2.6.0=py_1
tensorboard-data-server=0.6.0=py39hca6d32c_0
tensorboard-plugin-wit=1.6.0=py_0
tensorboardx=2.4.1=pypi_0
termcolor=1.1.0=py39h06a4308_1
terminado=0.13.1=pypi_0
testpath=0.5.0=pypi_0
texttable=1.6.4=pyhd3eb1b0_0
thinc=8.0.13=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h1ccaba5_0
tokenizers=0.10.3=py39hd6d55de_1
toml=0.10.2=pypi_0
tomli=2.0.0=pypi_0
torchaudio=0.10.0=py39_cu102
torchmetrics=0.7.2=pyhd8ed1ab_0
torchvision=0.11.1=py39_cu102
tornado=6.1=py39h27cfd23_0
tqdm=4.62.3=pyhd3eb1b0_1
traitlets=5.1.1=pypi_0
transformers=4.12.2=pyhd8ed1ab_0
typer=0.4.0=pypi_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
urllib3=1.26.7=pyhd3eb1b0_0
wandb=0.12.10=pypi_0
wasabi=0.9.0=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=2.0.2=pyhd3eb1b0_0
wheel=0.37.0=pyhd3eb1b0_1
widgetsnbextension=3.5.2=pypi_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yarl=1.6.3=py39h27cfd23_0
yaspin=2.1.0=pypi_0
zc-lockfile=2.0=pypi_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
